src/requirements.txt
pyTelegramBotAPI==4.23.0
python-telegram-bot==21.6
telebot==0.0.5
gitdb==4.0.11
GitPython==3.1.43
requests==2.25.1
